Moved upload action to components folder
