+++ frontend/src/components/UploadAction/UploadAction.jsx
@@ -2,10 +2,10 @@ import React, { useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import classNames from 'classnames';
 
-import { uploadFile } from '../../../../../store/reducers/drive/drive.thunks';
-import { vibrate } from '../../../../../utils/vibration';
+import { uploadFile } from '../../store/reducers/drive/drive.thunks';
+import { vibrate } from '../../utils/vibration';
 import UploadLoader from './UploadLoader';
-import { ReactComponent as PlusIcon } from '../../../../../assets/plus.svg';
+import { ReactComponent as PlusIcon } from '../../assets/plus.svg';
 
 import styles from './UploadAction.module.scss';
